+++ CHANGELOG.md
@@ -11,6 +11,10 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 - [#1834] PlaceObject4 tags appear as Unresolved inside of DefineSprite
 - [#1839] Sprite frames exported incorrectly and repeating
+- [#1838] AS3 - Properly handling of long unsigned values, hex values, default uint values etc.
+
+### Changed
+- AS3 integer values are internally (e.g. in the lib) handled as java int type instead of long.
 
 ## [15.1.1] - 2022-07-03
 ### Added
@@ -2376,6 +2380,7 @@ All notable changes to this project will be documented in this file.
 [alpha 7]: https://github.com/jindrapetrik/jpexs-decompiler/releases/tag/alpha7
 [#1834]: https://www.free-decompiler.com/flash/issues/1834
 [#1839]: https://www.free-decompiler.com/flash/issues/1839
+[#1838]: https://www.free-decompiler.com/flash/issues/1838
 [#270]: https://www.free-decompiler.com/flash/issues/270
 [#1718]: https://www.free-decompiler.com/flash/issues/1718
 [#1801]: https://www.free-decompiler.com/flash/issues/1801